requirements.txt
opencv-python
pytesseract
numpy
base64
python-dotenv
fireworks-client